feat(dialog): Add dialog to create elements
I added all ViewModels and Views for the Add Button. When we click on an item of the MenuItem in AddButton, we open a dialog with the corresponding view. Example : If we want to add a BankAccount, we click on the item, then the command is triggered with the corresponding View in the command parameters. In the MainWindowViewModel we have a method to open the dialog with the ViewModel get from the command parameter.