Feat(Add): Popup to add elements (#8)
* feat(dialog): Add dialog to create elements

I added all ViewModels and Views for the Add Button. When we click on an item of the MenuItem in AddButton, we open a dialog with the corresponding view. Example : If we want to add a BankAccount, we click on the item, then the command is triggered with the corresponding View in the command parameters. In the MainWindowViewModel we have a method to open the dialog with the ViewModel get from the command parameter.

* wip(add): Bank account wip (waiting for db svc)

* wip(add): Incomes wip (waiting for db svc)

* refacto(add): Change logic w. new svc + fix token

* feat(add): Add incomes dialog

* feat(add): Add saving projects dialog

* feat(add): Add saving dialog

* feat(add): Add spending dialog

* refacto(add): Incomes and Spending

Add nullable amount. Clean code. Change the UI for Incomes. Refacto AddIncomingViewModel

* fix(add): nullable float instead of float

* feat(style): Try to add style......

* chore(add): Clean code